Add new API request /api/v1/source/CHAPTER/SECTION which returns the source code for that section.
The tour controller now fetches the source through this API call. That way the
source code is always correctly escaped and we just set the initial source
in the controller init function. This eliminates the ugly source code setting
within the DIET template. And it later on enabled restoring user's changed
source editor content.