Added new 'docker' excution driver which runs 'stonemaster/dlang-tour-rdmd' for compiling and running.
This execution driver passes source code editor to an instance of that Docker image
which compiles the source and runs the program. This is done inside a Docker container
and should be a safe sandbox. The Docker container of the Dlang-Tour itself has been
adapted to use the Docker daemon on the host system to start those sandboxes. That
way the tour can reside inside Docker but start Docker containers on the host system.
Limits can be configured for the docker execution provider.

The dpaster exec provider has been removed for now.

Added systemd unit file for quickly deploying the dlang-tour using Docker.