Implement #57.
Sanity-checking can be done at startup of tour using the --sanitycheck flag.
All source code examples not markes as incomplete (title `{SourceCode:incomplete}`)
are checked for validity.

Also added travis CI so that is done before the final Docker image
is pushed.

Also changed travis-ci file to fail if Docker container is not built
or couldn't be pushed.